% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates_function.R
\name{agg_preprocess}
\alias{agg_preprocess}
\title{agg_preprocess}
\usage{
agg_preprocess(df)
}
\arguments{
\item{df}{a data frame with a key / value columns.}
}
\value{
A data frame.
}
\description{
\code{agg_preprocess} converts several country data variables to boolean in
order to caclulate the percentage of countries meeting a condition in the
aggregates tables
}
\seealso{
Other clean: 
\code{\link{bullying_clean}()},
\code{\link{chores_clean}()},
\code{\link{format_long}()},
\code{\link{format_wide}()},
\code{\link{gcpea_clean}()},
\code{\link{ict_skills_clean}()},
\code{\link{oecd_clean}()},
\code{\link{uis_clean}()},
\code{\link{un_aids_clean}()},
\code{\link{unicef_ecce_clean}()},
\code{\link{unicef_wash_clean}()},
\code{\link{wb_clean}()},
\code{\link{weights_clean}()}
}
\concept{clean}
