% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates_function.R
\name{aggregates}
\alias{aggregates}
\title{aggregates}
\usage{
aggregates(df)
}
\arguments{
\item{df}{data frame in the same format as \code{country_data2}}
}
\description{
\code{aggregates} is a function to calculate income group and regional aggregates.
}
\details{
Calculates, medians, weighted means, sums, and 'percentage true' aggregations (as
specified in the indicators dataframe). Takes \code{country_data2} (produced as part
of the \code{gemrtables} function) as an argument.
}
\seealso{
\code{\link{gemrtables}}

Other summarise: \code{\link{parity_adj}}
}
\concept{summarise}
