% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates_function.R
\name{aggregates_values}
\alias{aggregates_values}
\title{aggregates_values}
\usage{
aggregates_values(df)
}
\arguments{
\item{df}{a data frame with a key / value columns.}
}
\value{
A data frame.
}
\description{
\code{aggregates_values} Is a wrapper function for \code{compute_aggregate_values}.
Aggregates values by medians at the region level.
}
\seealso{
Other summarise: \code{\link{aggregates}},
  \code{\link{compute_aggregate_values}},
  \code{\link{compute_aggregate}},
  \code{\link{parity_adj}},
  \code{\link{parity_indices_region}}
}
\concept{summarise}
