% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{bullying_clean}
\alias{bullying_clean}
\title{bullying_clean}
\usage{
bullying_clean(df)
}
\arguments{
\item{df}{a data frame from Innocenti}
}
\description{
\code{bullying_clean} is a function to clean the Innocenti bullying data on
google drive.
}
\details{
Information on survey used per observation is removed.
}
\seealso{
\code{\link{other}}

Other clean: \code{\link{agg_preprocess}},
  \code{\link{chores_clean}}, \code{\link{format_long}},
  \code{\link{format_wide}}, \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{oecd_clean}},
  \code{\link{uis_clean}}, \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
