% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cedar_import.R
\name{cedar}
\alias{cedar}
\title{cedar}
\usage{
cedar()
}
\description{
\code{uis} is a function to import and clean cedar data
}
\details{
Defines relvent tables and filters for cedar sql tables and applies the `cedar_clean` function
}
\seealso{
\code{\link{cedar_clean}}

Other import/clean function: \code{\link{inds}},
  \code{\link{other}}, \code{\link{region_groups}},
  \code{\link{uis}}, \code{\link{weights}}
}
\concept{import/clean function}
