% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{cedar_clean}
\alias{cedar_clean}
\title{cedar_clean}
\usage{
cedar_clean(df)
}
\description{
\code{cedar_clean} is a function to clean data imported by \code{read_cedar}.
}
\details{
Cleans data imported from cedar database and computes several variables
(CR.1.GPIA, CR.1.LPIA, CR.1.WPIA, CR.2.GPIA, CR.2.LPIA, CR.2.WPIA, CR.3.GPIA,
CR.3.LPIA, CR.3.WPIA, chores.28plus.12t14.GPIA)
}
\seealso{
\code{\link{read_cedar}}

Other clean: \code{\link{agg_preprocess}},
  \code{\link{bullying_clean}}, \code{\link{chores_clean}},
  \code{\link{format_long}}, \code{\link{format_wide}},
  \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{oecd_clean}},
  \code{\link{uis_clean}}, \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
