% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{chores_clean}
\alias{chores_clean}
\title{chores_clean}
\usage{
chores_clean(df)
}
\arguments{
\item{df}{a data frame}
}
\description{
\code{chores_clean} is a function to clean the child chores data on
google drive.
}
\seealso{
\code{\link{other}}

Other clean: \code{\link{agg_preprocess}},
  \code{\link{bullying_clean}}, \code{\link{format_long}},
  \code{\link{format_wide}}, \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{oecd_clean}},
  \code{\link{uis_clean}}, \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
