% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates_function.R
\name{compute_aggregate_values}
\alias{compute_aggregate_values}
\title{compute_aggregate_values}
\usage{
compute_aggregate_values(df, region, entity)
}
\arguments{
\item{df}{a data frame with a key / value columns.}

\item{region}{aggregate grouping on which to compute (corresponds to the
relevent column in `.gemrtables.pkg.env$indicators`.}

\item{entity}{type of aggregate (character)}
}
\value{
A data frame.
}
\description{
\code{compute_aggregate_values} computes medians on country_data
for a specified aggregate grouping.
}
\seealso{
Other summarise: \code{\link{aggregates_values}},
  \code{\link{aggregates}},
  \code{\link{compute_aggregate}},
  \code{\link{parity_adj}},
  \code{\link{parity_indices_region}}
}
\concept{summarise}
