% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{eurostat_clean}
\alias{eurostat_clean}
\title{eurostat_clean}
\usage{
eurostat_clean(df)
}
\description{
\code{eurostat_clean} is a function to clean data from eurostat API queries
defined in \code{other}.
}
\seealso{
\code{\link{other}}
}
\concept{clean function}
