% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{format_long}
\alias{format_long}
\title{format_long}
\usage{
format_long()
}
\description{
\code{format_long} is a function to format stat table data to 'long' format.
}
\seealso{
Other clean: \code{\link{agg_preprocess}},
  \code{\link{bullying_clean}}, \code{\link{chores_clean}},
  \code{\link{format_wide}}, \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{oecd_clean}},
  \code{\link{uis_clean}}, \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
