% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{format_wide}
\alias{format_wide}
\title{format_wide}
\usage{
format_wide(df)
}
\description{
\code{format_wide} is a function to format stat table data to 'wide' format.
}
\details{
Rounds values, converts binary values to 'Yes/No'; converts value to unicode
with subscript flags; converts to list of dataframes for export to xlsx.
}
\seealso{
Other clean: 
\code{\link{agg_preprocess}()},
\code{\link{bullying_clean}()},
\code{\link{chores_clean}()},
\code{\link{format_long}()},
\code{\link{gcpea_clean}()},
\code{\link{ict_skills_clean}()},
\code{\link{oecd_clean}()},
\code{\link{uis_clean}()},
\code{\link{un_aids_clean}()},
\code{\link{unicef_ecce_clean}()},
\code{\link{unicef_wash_clean}()},
\code{\link{wb_clean}()},
\code{\link{weights_clean}()}
}
\concept{clean}
