% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{gcpea_clean}
\alias{gcpea_clean}
\title{gcpea_clean}
\usage{
gcpea_clean(df)
}
\description{
\code{gcpea_clean} is a function to clean data from Global Campaign to
protect education from attack (GCPEA) on google drive.
}
\details{
'm' val_status indicates observation comes from a multi-year period and does
not reflect the number of attacks on education in a single year period).
}
\seealso{
\code{\link{other}}

Other clean: 
\code{\link{agg_preprocess}()},
\code{\link{bullying_clean}()},
\code{\link{chores_clean}()},
\code{\link{format_long}()},
\code{\link{format_wide}()},
\code{\link{ict_skills_clean}()},
\code{\link{oecd_clean}()},
\code{\link{uis_clean}()},
\code{\link{un_aids_clean}()},
\code{\link{unicef_ecce_clean}()},
\code{\link{unicef_wash_clean}()},
\code{\link{wb_clean}()},
\code{\link{weights_clean}()}
}
\concept{clean}
