% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemrtables_function.R
\name{gemrtables}
\alias{gemrtables}
\title{gemrtables}
\usage{
gemrtables(region = "SDG.region", ref_year, export = FALSE, path, key,
  password)
}
\arguments{
\item{region}{regional grouping on which to compute aggregate values. Options
are `GEMR.region`, `UIS.region` or `SDG.region`. Defaults to `SDG.region`.}

\item{ref_year}{A four digit numeric of the reference year. Default is two
years prior to the current year.}

\item{export}{If `TRUE` returns an xlsx workbook in 'wide' format, with
  seperate worksheets per table.
If `FALSE` (the default) returns an data frame in 'long' format.}

\item{path}{File path to write xlsx workbook (character). Overwrites existing
file.}

\item{key}{UIS api subcription key.}

\item{password}{password to cedar database.}
}
\value{
A data frame or an xlsx workbook.
}
\description{
\code{gemrtables} is the main function to generate the UNESCO Global
Education Monitoring (GEM) Report Statistical tables, returning a dataframe
or an xlsx workbook
}
\examples{
gemrtables(ref_year = 2016, export = TRUE, path = x, key = y, password = z)

df <- gemrtables(region = "UIS.region", ref_year = 2016, export = FALSE, key = y, password = z)
}
