% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uis_alternative_access.R
\name{get_uis_altcodes}
\alias{get_uis_altcodes}
\title{get_uis_altcodes}
\usage{
get_uis_altcodes()
}
\description{
Fallback access to data.uis.unesco.org SDMX interface, for aggregates missing from api.uis.unesco.org
}
