% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{inds}
\alias{inds}
\title{inds}
\usage{
inds()
}
\description{
\code{other} is a function to import statistical table indicators.
}
\details{
This dataframe defines indicators to be included in the statistical tables.
It's purpose is to filter, rename and order variables, and define methods of
aggregation.
}
\seealso{
Other import/clean: \code{\link{cedar}},
  \code{\link{other}}, \code{\link{region_groups}},
  \code{\link{uis}}, \code{\link{weights}}
}
\concept{import/clean}
