% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{oecd_clean}
\alias{oecd_clean}
\title{oecd_clean}
\usage{
oecd_clean(df)
}
\description{
\code{oecd_clean} is a function to clean data from OECD API queries defined
in \code{other}.
}
\details{
Cleans data from OECD api queries and computes one variable.
(sal.rel.2t3,)
}
\seealso{
\code{\link{other}}

Other clean: \code{\link{agg_preprocess}},
  \code{\link{bullying_clean}}, \code{\link{cedar_clean}},
  \code{\link{chores_clean}}, \code{\link{format_wide}},
  \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{uis_clean}},
  \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
