% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_import.R
\name{other}
\alias{other}
\title{other}
\usage{
other()
}
\description{
\code{other} is a function to import and clean miscellaneous data.
}
\details{
Defines  queries to the Eurostat, OECD and World Bank APIs, miscellaneous
spreadsheets, and applies respective `clean` functions.
}
\seealso{
#'\code{\link{eurostat_clean}},  '\code{\link{gcpea_clean}},  '\code{\link{un_aids_clean}},
 '\code{\link{unicef_ecce_clean}},  \code{\link{unicef_wash_clean}},  \code{\link{wb_clean}},

Other import/clean: \code{\link{cedar}},
  \code{\link{inds}}, \code{\link{region_groups2}},
  \code{\link{region_groups}}, \code{\link{uis}},
  \code{\link{weights}}
}
\concept{import/clean}
