% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{parity_adj}
\alias{parity_adj}
\title{parity_adj}
\usage{
parity_adj(df, col, a, b, varname, val_status = FALSE)
}
\arguments{
\item{df}{a data frame with a key / value columns.}

\item{col}{indicator key column.}

\item{a}{indicator for 'disadvantaged' group (numerator).}

\item{b}{indicator for 'advantaged' group  (denominator).}

\item{varname}{name for calculated indice (character).}

\item{val_status}{For use with data with flags for estimated observations. If
`TRUE` will calculate flag for indice (requires flag column to be named
`val_status` and estimates lablled as `E`.}
}
\value{
A data frame.
}
\description{
\code{parity_adj} calculates adjusted parity ratios.
}
\examples{

parity_adj(df, indicator, ger.female, ger.male, "ger.gpia", val_status = FALSE)

parity_indices <- list(df = list("df","df"),
col = list("ind",  "ind"),
a = list("adult.profiliteracy.f","adult.profinumeracy.f"),
b = list("adult.profiliteracy.m", "adult.profinumeracy.m"),
varname = list("adult.profiliteracy.gpia", "adult.profinumeracy.gpia")) \%>\%
pmap(parity_adj) \%>\%
}
\concept{summarise}
