% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_cedar}
\alias{read_cedar}
\title{read_cedar}
\usage{
read_cedar(sc, level = 1, table, ind, password)
}
\arguments{
\item{sc}{SQL table on which to filter observations. Current choices are
`wide_dimension` for wide data and `other_dimension` for GEM 2030 data.}

\item{level}{Disaggregations on which to filter `wide_dimension` table
(connect to sql database for further details). Defaults to 1 (total
country).}

\item{table}{Table of variables on which to join (character).}

\item{ind}{Character vector of variables to select.}

\item{password}{Password to connect to cedar SQL database.}
}
\value{
A data frame.
}
\description{
\code{read_cedar} is a function to join, filter and import data from GEMR cedar SQL database
}
\examples{
read_cedar(sc = "wide_dimension", level = 1, table = "wide_1", ind =
c("trans_prim_m", "comp_prim_v2_m"), password = x)

list(sc = list("wide_dimension",
"wide_dimension",
"other_dimension",
"other_dimension",
"other_dimension",
"other_dimension"),
level = list(1,
            c(13:15, 31),
            NA,
            NA,
            NA,
            NA),
table = list("wide_1",
            "wide_1",
            "other_school_readiness",
            "other_facilities",
            "other_social_norms",
            "other_curriculum"),
ind = list(c("trans_prim_m", "comp_prim_v2_m", "comp_lowsec_v2_m", "comp_upsec_v2_m"),
          c("comp_prim_v2_m", "comp_lowsec_v2_m", "comp_upsec_v2_m"),
          c("u5_posit_home_learn", "u5_child_book", "school_child_track"),
          c("stu_exper_bully_13_17", "stu_exper_violence_13_17"),
          "child_chores_more_28_12_14",
          c("esd_gced_curr_ge", "esd_gced_curr_hr", "esd_gced_glo_cit", "esd_gced_sus_dev"))) \%>\%
 pmap(read_cedar, password = x)
}
\seealso{
codebook for WIDE indicators:
  \url{https://drive.google.com/file/d/0B5qc8r9eSwe4LUhWajNwTS10TDA/view}.
  Codebook for GEM2030 indicators:
  \url{https://sites.google.com/view/codebooks/home/gem-2030}

Other import functions: \code{\link{read_urls}}
}
\concept{import functions}
