% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_functions.R
\name{read_urls}
\alias{read_urls}
\title{read_urls}
\usage{
read_urls(urls, key = NULL, bind = TRUE)
}
\arguments{
\item{urls}{SDMX url query or list of SDMX url queries (character).}

\item{key}{UIS subcription key. Required if `urls` contains UIS queries.
Default is NULL.}

\item{bind}{If `TRUE` returns a list a data frames for each query. If `FALSE` binds
queries into a single data frame.}
}
\value{
A data frame.
}
\description{
\code{read_urls} parses SDMX rest url(s) to data frames
}
\details{
Wrapper for \code{\link[rsdmx]{readSDMX}}
}
\examples{
read_urls("https://api.uis.unesco.org/sdmx/data/UNESCO,EDU_NON_FINANCE,2.0/COMP_EDU.YR.L1..................?format=sdmx-compact-2.1&lastNObservations=1&subscription-key=",
key = x)

read_urls("http://ec.europa.eu/eurostat/SDMX/diss-web/rest/data/trng_aes_100/.T.FE_NFE.PC../?startperiod=2010&endPeriod=2050")

list(
"https://stats.oecd.org/restsdmx/sdmx.ashx/GetData/CRS1/20005..11420.100.100.D.112.E02+E01/all?startTime=2016&endTime=2050",
"https://stats.oecd.org/restsdmx/sdmx.ashx/GetData/EAG_TS_ACT/..L0+L1+L2_C4+L3_C4.Y25T64.T.RATIO_ACTL_TER/all?",
"https://stats.oecd.org/restsdmx/sdmx.ashx/GetData/EDU_PERS_INST/.T.INST_T.T.L2+L3.T.TEACH.PER/all?startTime=2010&endTime=2050") \%>\%
 read_urls(bind = FALSE))

list(
"https://api.uis.unesco.org/sdmx/data/UNESCO,EDU_NON_FINANCE,2.0/COMP_EDU.YR.L1..................?format=sdmx-compact-2.1&lastNObservations=1&subscription-key=",
"https://stats.oecd.org/restsdmx/sdmx.ashx/GetData/EAG_TS_ACT/..L0+L1+L2_C4+L3_C4.Y25T64.T.RATIO_ACTL_TER/all?",
"https://stats.oecd.org/restsdmx/sdmx.ashx/GetData/EDU_PERS_INST/.T.INST_T.T.L2+L3.T.TEACH.PER/all?startTime=2010&endTime=2050") \%>\%
 read_urls(key = x))
}
\seealso{
\code{\link[rsdmx]{readSDMX}}

Other import: \code{\link{read_cedar}}
}
\concept{import}
