% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{region_groups}
\alias{region_groups}
\title{region_groups}
\usage{
region_groups()
}
\description{
\code{region_groups} is a function to import country names and regional groupings.
}
\details{
This dataframe defines various country names, iso / OECD CRS codes, and
regional groupings on which to calculate aggregates.
}
\seealso{
Other import/clean: \code{\link{inds}},
  \code{\link{other}}, \code{\link{region_groups2}},
  \code{\link{uis}}, \code{\link{weights}}
}
\concept{import/clean}
