% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{region_groups2}
\alias{region_groups2}
\title{region_groups2}
\usage{
region_groups2()
}
\description{
\code{region_groups2} is a function import regional groupings, ordering, and REF_AREA api codes.
}
\seealso{
Other import/clean: \code{\link{inds}},
  \code{\link{other}}, \code{\link{region_groups}},
  \code{\link{uis}}, \code{\link{weights}}
}
\concept{import/clean}
