% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uis_import.R
\name{uis}
\alias{uis}
\title{uis}
\usage{
uis()
}
\description{
\code{uis} is a function to import and clean UIS data.
}
\details{
Defines SDMX queries to the UIS API and applies the `uis_clean` function
}
\seealso{
\code{\link{uis_clean}}

Other import/clean: \code{\link{inds}},
  \code{\link{other}}, \code{\link{region_groups2}},
  \code{\link{region_groups}}, \code{\link{weights}}
}
\concept{import/clean}
