% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{uis_clean}
\alias{uis_clean}
\title{uis_clean}
\usage{
uis_clean(df)
}
\description{
\code{uis_clean} is a function to clean UIS data.
}
\details{
Cleans data from UIS api queries and computes several variables
(admi.grade2or3prim, admi.endofprim, admi.endoflowersec, Comp.02, Free.02,
Comp.2t3, Free.2t3, Read.Primary.GPIA, Math.Primary.GPIA, Read.LowerSec.GPIA,
Math.LowerSec.GPIA, LR.Ag15t24.GPIA, LR.Ag15t99.GPIA, Read.Primary.WPIA,
Math.Primary.WPIA, Read.LowerSec.WPIA, Math.LowerSec.WPIA)
}
\seealso{
Other clean: \code{\link{agg_preprocess}},
  \code{\link{bullying_clean}}, \code{\link{cedar_clean}},
  \code{\link{chores_clean}}, \code{\link{format_long}},
  \code{\link{format_wide}}, \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{oecd_clean}},
  \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
