% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{un_aids_clean}
\alias{un_aids_clean}
\title{un_aids_clean}
\usage{
un_aids_clean(df)
}
\description{
\code{un_aids_clean} is a function to clean data from the UN AIDS spreadsheet
on google drive.
}
\details{
Information on survey used per observation is removed.
}
\seealso{
\code{\link{other}}

Other clean: \code{\link{cedar_clean}},
  \code{\link{format_wide}}, \code{\link{gcpea_clean}},
  \code{\link{oecd_clean}}, \code{\link{uis_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
