% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{unicef_ecce_clean}
\alias{unicef_ecce_clean}
\title{unicef_ecce_clean}
\usage{
unicef_ecce_clean(df, ind, source)
}
\arguments{
\item{df}{a datadrame in the same format as
\url{https://drive.google.com/uc?export=download&id=1JzOo7rt8O3ZE9eSAL3v1jKexLwbICtM3}}

\item{ind}{name of the indicator of the `value` column in the csv file
(character).}

\item{source}{name of the source of the data (i.e. UNICEF)}
}
\description{
\code{unicef_ecce_clean} is a function to clean UNICEF ECCE survey data on
google drive.
}
\details{
Information on survey used per observation is removed.
}
\seealso{
\code{\link{other}}

Other clean: \code{\link{agg_preprocess}},
  \code{\link{bullying_clean}}, \code{\link{cedar_clean}},
  \code{\link{chores_clean}}, \code{\link{format_wide}},
  \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{oecd_clean}},
  \code{\link{uis_clean}}, \code{\link{un_aids_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
