% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{unicef_wash_clean}
\alias{unicef_wash_clean}
\title{unicef_wash_clean}
\usage{
unicef_wash_clean(df)
}
\description{
\code{unicef_wash_clean} is a function to clean data from the UNICEF WASH
file on google drive.
}
\seealso{
\code{\link{other}}

Other clean function: \code{\link{cedar_clean}},
  \code{\link{eurostat_clean}}, \code{\link{format_wide}},
  \code{\link{gcpea_clean}}, \code{\link{oecd_clean}},
  \code{\link{uis_clean}}, \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean function}
