% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_files.R
\name{weights}
\alias{weights}
\title{weights}
\usage{
weights()
}
\description{
\code{weights} is a function to import and clean weights data
}
\details{
Defines SDMX queries to the UIS / UN APIs and applies the `weights_clean`
function
}
\seealso{
\code{\link{weights_clean}}

Other import/clean function: \code{\link{cedar}},
  \code{\link{inds}}, \code{\link{other}},
  \code{\link{region_groups}}, \code{\link{uis}}
}
\concept{import/clean function}
