% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{weights_clean}
\alias{weights_clean}
\title{weights_clean}
\usage{
weights_clean(df)
}
\description{
\code{weights_clean} is a function to clean weights data imported from SDMX
queries in `weights` function.
}
\details{
Several variables are calculated (Y25T65, Y_GE25, Y15T64, L1_GLAST_Q1_F,
L1_GLAST_Q1_M, L2_GLAST_Q1_F, L2_GLAST_Q1_M, L3_Q1_F = L3_F, L3_Q1_M = L3_M).
}
\seealso{
\code{\link{weights}}

Other clean: \code{\link{agg_preprocess}},
  \code{\link{bullying_clean}}, \code{\link{cedar_clean}},
  \code{\link{chores_clean}}, \code{\link{format_long}},
  \code{\link{format_wide}}, \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{oecd_clean}},
  \code{\link{uis_clean}}, \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}}
}
\concept{clean}
