% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregates_function.R
\name{aggregates}
\alias{aggregates}
\title{aggregates}
\usage{
aggregates(df)
}
\arguments{
\item{df}{a data frame with a key / value columns.}
}
\value{
A data frame.
}
\description{
\code{aggregates} Is a wrapper function for \code{compute_aggregate}.
Aggregates are computed for each type of grouping and binded. Redundant
aggregations are dropped through a join.
}
\seealso{
Other summarise: \code{\link{compute_aggregate}},
  \code{\link{parity_adj}}
}
\concept{summarise}
