% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{eurostat_clean}
\alias{eurostat_clean}
\title{eurostat_clean}
\usage{
eurostat_clean(df)
}
\description{
\code{eurostat_clean} is a function to clean data from eurostat API queries
defined in \code{other}.
}
\seealso{
\code{\link{other}}

Other clean function: \code{\link{cedar_clean}},
  \code{\link{format_wide}}, \code{\link{gcpea_clean}},
  \code{\link{oecd_clean}}, \code{\link{uis_clean}},
  \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}}, \code{\link{wb_clean}},
  \code{\link{weights_clean}}
}
\concept{clean function}
