% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemrtables.R, R/gemrtables_function.R
\docType{package}
\name{gemrtables}
\alias{gemrtables}
\alias{gemrtables-package}
\title{gemrtables: A package for generating the Global Education Monitoring (GEM)
Report statistical tables.}
\usage{
gemrtables(region = "SDG.region", ref_year, export = TRUE, path, key,
  pc_flag_cut = 66, pc_comp_cut2 = 33, level_country = TRUE,
  drake = TRUE, removeCache)
}
\arguments{
\item{region}{regional grouping on which to compute aggregate values. Options
are `GEMR.region`, `UIS.region` or `SDG.region`. Defaults to `SDG.region`.}

\item{ref_year}{A four digit numeric of the reference year. Default is two
years prior to the current year.}

\item{export}{If `TRUE` returns an xlsx workbook in 'wide' format, with
  separate worksheets per table.
If `FALSE` (the default) returns a data frame in 'long' format.}

\item{path}{File path to write xlsx workbook (character). Overwrites existing
file.}

\item{key}{UIS api subscription key.}

\item{pc_flag_cut}{threshold}

\item{pc_comp_cut2}{threshold}

\item{level_country}{If it is TRUE, parity indices are calculated at the country level}

\item{drake}{If it is TRUE, the drake plan is run instead load or generate data}

\item{removeCache}{Character vector of unprocessed dataframes. Options are c("uis_up", "wb_up", "eurostat_up", "oecd_up")}
}
\value{
A data frame or an xlsx workbook.
}
\description{
\code{\link{gemrtables}} is the primary function for this package, generating
the statistical tables in a 'wide' data format in excel, or a 'long' format
data frame.

\code{gemrtables} is the main function to generate the UNESCO Global
Education Monitoring (GEM) Report Statistical tables, returning a dataframe
or an xlsx workbook
}
\details{
In addition, the gemrtables package provides three categories of functions:
import, import/clean, clean and summarise.
}
\section{import functions}{

The import functions read data from SDMX url queries or the GEM cedar SQL
database.
}

\section{clean functions}{

Clean functions sort data into a standard 'long' format with six columns:

\enumerate{
  \item iso2c: ISO 3166-1 alpha-2 country code
  \item year: year of observation
  \item ind: standardised indicator name, matching with that in the inds()
  data frame (for indicators to be included in the stat tables)
  \item value: observation value
  \item val_status: observation status
  \item source of the observation (matching to the inds() data frame)
}
In addition several clean function compute new observations.
}

\section{summarise functions}{

summarise functions compute new observations -aggregates and parity indices
}

\section{import/clean functions}{

These functions largely serve as a wrapper for import and clean functions.
}

\examples{
gemrtables(ref_year = 2016, export = TRUE, path = x, key = y)

df <- gemrtables(region = "UIS.region", ref_year = 2016, export = FALSE, key = y)
}
