% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{parity_indices_region}
\alias{parity_indices_region}
\title{parity_indices_region}
\usage{
parity_indices_region()
}
\value{
A data frame.
}
\description{
\code{parity_indices_region} calculates adjusted parity ratios at the region level.
}
\seealso{
Other summarise: \code{\link{aggregates_values}},
  \code{\link{aggregates}},
  \code{\link{compute_aggregate_values}},
  \code{\link{compute_aggregate}}, \code{\link{parity_adj}}
}
\concept{summarise}
