% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_functions.R
\name{wb_clean}
\alias{wb_clean}
\title{wb_clean}
\usage{
wb_clean(df)
}
\arguments{
\item{df}{a data frame from WB}
}
\description{
\code{wb_clean} is a function to clean data imported by \code{[wbstats]{wb}} within \code{\link{other}} .
}
\details{
Cleans data imported by \code{[wbstats]{wb}} and computes two variables
(adult.profiliteracy.gpia, adult.profinumeracy.gpia)
}
\seealso{
\code{\link[wbstats]{wb}}, \code{\link{other}}

Other clean: \code{\link{agg_preprocess}},
  \code{\link{bullying_clean}}, \code{\link{chores_clean}},
  \code{\link{format_long}}, \code{\link{format_wide}},
  \code{\link{gcpea_clean}},
  \code{\link{ict_skills_clean}}, \code{\link{oecd_clean}},
  \code{\link{uis_clean}}, \code{\link{un_aids_clean}},
  \code{\link{unicef_ecce_clean}},
  \code{\link{unicef_wash_clean}},
  \code{\link{weights_clean}}
}
\concept{clean}
